pyproject.toml
[tool.poetry]
name = "SeSePerson"
version = "0.0.0"
description = "A highly intelligent AI"
authors = ["coming <kangming06@hotmail.com>"]
readme = "README.md"
license = "GPL-3.0"
repository="https://github.com/wellcoming/SeSePerson"


[[tool.poetry.source]]
name = "tsinghua-mirror"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"


[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = {extras = ["fastapi"], version = "^2.1.3"}
nonebot-adapter-onebot = "^2.3.1"
tortoise-orm = "^0.20.0"
langchain-core = {git = "https://github.com/wellcoming/langchain.git", subdirectory = "libs/core"}
langchain = "^0.1.17"
langchain-openai = "^0.1.4"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
